evm/eof: rename eofContainer.ts to container.ts
